python/requirements.txt
# Requires Python 3.11 on Windows (pre-built wheels available)
bless>=0.2.5
bleak>=0.21.0,<0.22.0
bleak-winrt>=1.2.0 ; sys_platform == 'win32'
winrt-Windows.Devices.Bluetooth ; sys_platform == 'win32'
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile ; sys_platform == 'win32'
winrt-Windows.Devices.Bluetooth.Advertisement ; sys_platform == 'win32'
winrt-Windows.Storage.Streams ; sys_platform == 'win32'
winrt-Windows.Foundation ; sys_platform == 'win32'
winrt-Windows.Foundation.Collections ; sys_platform == 'win32'
pyinstaller>=6.0.0
pysetupdi @ git+https://github.com/gwangyi/pysetupdi.git
